python/requirements.txt
torch==2.1.0
torchvision==0.16.0
Pillow==9.3.0
numpy==1.23.4
opencv-python==4.6.0.66
mediapipe==0.9.1
Flask==2.3.2
Flask-Cors==3.1.1
google-generativeai==0.0.3
selenium==4.9.0
webdriver-manager==4.9.1
requests==2.28.2
